Mical theme is simple Ui based responsive website which is created by using HTML% and CSS3
t is My first website UI design which is inspired by a demo website for Car services name 'Mical' , It is
a responsive but not fully functioned website to design the UI of 'Mical' I used HTML5 and CSS3 only.
A part from that i applied font awesome and animation libraries.